// src/types/kyc.ts
export interface UserData {
  fullName?: string;
  cpf?: string;
  birthDate?: string;
  referenceImg?: string | File ; 
  fatherUrl?: string;
  front?: string | File;        
  back?: string | File;         
}
// ...
export interface PepData {
  name: string;
  document: string;
  typeIndicator: string;
  pepIndicator: string;
  dateBirth: string;
  hasMandatesPermission: boolean;
  hasAssociatePermission: boolean;
  hasRelativePermission: boolean;
  messageMandatesPermission: string;
  messageAssociatePermission: string;
  messageRelativePermission: string;
  mandates: PepMandate[];
  associates: PepAssociate[];
  relatives: PepRelative[];
}

export interface Enrichment {
  document: string;
  pep: PepData;
}

export interface ResponseWebService{
  approved:Boolean 
  message:String 
  enrichments?: Enrichment[];

}